% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_endpoints.R
\name{add_forced_group_levels}
\alias{add_forced_group_levels}
\title{Add forced group levels}
\usage{
add_forced_group_levels(combos_all, forced_group_levels)
}
\arguments{
\item{combos_all}{A data.table containing all combinations of group levels found in the analysis data.}

\item{forced_group_levels}{A one column data.table containing a required set of group levels of a grouping variable.}
}
\value{
A data.table containing all combinations of group levels exapnded with the forced grouping levels.
}
\description{
Expand the set of unique group levels of one grouping variables in a table containing all combinations of one or more grouping variables.
}
