% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_db_data.R, R/mk_filtered_endpoint_dt.R
\name{apply_dt_filter}
\alias{apply_dt_filter}
\title{Apply filtering logic to a data.table}
\usage{
apply_dt_filter(adam_dt, filter_string, type = c("filter", "flag"))

apply_dt_filter(adam_dt, filter_string, type = c("filter", "flag"))
}
\arguments{
\item{adam_dt}{data.table::data.table}

\item{filter_string}{character}

\item{type}{character}
}
\value{
A \code{data.table} that has been filtered according to the specified
logic, or with an added flag column.

data.table::data.table
}
\description{
This utility function applies a specified filter to a \code{data.table}. The
filter can be applied to either subset the data (type = "filter") or to
create a new flag column within the data (type = "flag").
}
