% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_stats.R
\name{apply_stats}
\alias{apply_stats}
\title{Apply Statistical Methods}
\usage{
apply_stats(
  ep,
  analysis_data_container,
  type = c("stat_by_strata_by_trt", "stat_by_strata_across_trt",
    "stat_across_strata_across_trt")
)
}
\arguments{
\item{ep}{A \code{data.table} containing prepared endpoint data for statistical
analysis, typically the output from \code{prepare_for_stats}.}

\item{analysis_data_container}{data.table containing the analysis data.
functions.}

\item{type}{The type of statistical function. Can be one of
"stat_by_strata_by_trt", "stat_by_strata_across_trt", or "stat_across_strata_across_trt"}
}
\value{
A \code{data.table} with statistical results appended.
}
\description{
Applies user-supplied statistical functions to the prepared
endpoint data. The results are stored in a \code{data.table} that is nested in
the row of the endpoint definition object.
}
