% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_stats.R
\name{define_expansion_cell_from_data}
\alias{define_expansion_cell_from_data}
\title{Create Expansion Cell Containing a Data Table Based on Strata}
\usage{
define_expansion_cell_from_data(row, grouping_cols, data_col, col_prefix)
}
\arguments{
\item{row}{A single row from a data table containing endpoint data.}

\item{grouping_cols}{A character vector specifying the columns that point to
the grouping variables in the data.}

\item{data_col}{The name of the column that points to the ADaM dataset.}

\item{col_prefix}{A prefix for the resulting columns, such as "prefix_filter"
and "prefix_metadata".}
}
\value{
A nested \code{data.table} (double listed) to be inserted into a data table
cell.
}
\description{
Defines an expansion cell for a \code{data.table} row based on the
the strata. Creates a \code{data.table} with metadata and filter conditions for
each combination of strata levels.
}
