% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_stats.R
\name{expand_ep_for_stats}
\alias{expand_ep_for_stats}
\title{Expand Endpoint Data for Statistics}
\usage{
expand_ep_for_stats(
  ep,
  grouping_cols,
  analysis_data_container,
  data_col,
  id_col,
  col_prefix
)
}
\arguments{
\item{ep}{A \code{data.table} containing endpoint data to be expanded.}

\item{grouping_cols}{A character vector specifying the columns used for
grouping in the expansion.}

\item{data_col}{The name of the column in \code{ep} that contains the ADaM
dataset.}

\item{id_col}{The name of the column in \code{ep} that contains the unique
identifier for the strata.}

\item{col_prefix}{A prefix for the resulting columns, defaults to "stat".}
}
\value{
A \code{data.table} with expanded endpoints prepared for statistical
analysis.
}
\description{
Expands endpoint datatable for statistical analysis based on
specified grouping columns. Creates specifications for expansion and
performs the expansion to prepare the data for statistical functions.
}
