% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_db_data.R
\name{fetch_db_data}
\alias{fetch_db_data}
\title{Fetch ADaM tables from a source based on endpoint
specifications}
\usage{
fetch_db_data(study_metadata, fn_dt, env = parent.frame())
}
\arguments{
\item{study_metadata}{Study metadata that is passed to the \verb{mk_adam*} functions.}

\item{fn_dt}{A \code{data.table} containing the parsed user-supplied functions.}

\item{env}{Advanced parameter for specifying the evaluation environment, defaults to \code{parent.frame()}.}
}
\value{
A list of ADaM tables with additional columns from \code{fn_dt} including function type, hash, name,
character representation of the call, and the callable itself.
}
\description{
This function reads the set of ADaM tables required to produce the specified
endpoints from the input. It uses the user-supplied preprocessing functions
specified for each endpoint to identify and process the necessary ADaM
tables. It validates the preprocessing functions and applies them to the
study metadata to obtain the data tables. If any preprocessing functions
result in errors, the function stops and reports the errors.
}
