% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_db_data.R
\name{filter_adam_db}
\alias{filter_adam_db}
\title{Title}
\usage{
filter_adam_db(dat, pop_var, pop_value, custom_pop_filter)
}
\arguments{
\item{dat}{A data.table containing the analysis data.}

\item{pop_var}{Population flag variable.}

\item{pop_value}{A character containing the acceptance value of population
flag value.}

\item{custom_pop_filter}{A character string containing a custom population
filter.}
}
\value{
A data.table containing the filtered analysis data.
}
\description{
Title
}
