% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_db_data.R
\name{filter_db_data}
\alias{filter_db_data}
\title{Filter study data based on endpoint specifications}
\usage{
filter_db_data(ep, ep_fn_map, adam_db)
}
\arguments{
\item{ep}{A \code{data.table} containing endpoint definitions.}

\item{ep_fn_map}{A \code{data.table} mapping endpoint definitions to function
hashes and types.}

\item{adam_db}{A list of ADaM tables obtained from \code{fetch_db_data}.}
}
\value{
A \code{data.table} with filtered datasets for each endpoint in the study.
}
\description{
This function filters ADaM datasets according to the specifications for each
endpoint. It merges endpoint definitions with a mapping of endpoints to
function hashes and then with the actual ADaM datasets. Custom filtering
logic is applied to each dataset to produce the output dataset.
}
