% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_stats_results.R
\name{format_stats_results}
\alias{format_stats_results}
\title{Format results}
\usage{
format_stats_results(ep)
}
\arguments{
\item{ep}{A data.table containing the endpoint information.}
}
\value{
A data.table containing the unnested endpoint information.
}
\description{
Format results
}
