% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_userdef_fn_dt.R
\name{generate_function_table_row}
\alias{generate_function_table_row}
\title{Generate a row for the function table}
\usage{
generate_function_table_row(fn_type, fn, fn_name, fn_hash, env)
}
\arguments{
\item{fn_type}{The type of the function as a character string.}

\item{fn}{A list containing the function to be called.}

\item{fn_name}{The name of the function as a character string.}

\item{fn_hash}{The unique hash of the function as a character string.}

\item{env}{The environment in which to evaluate the function.}
}
\value{
A \code{data.table} row with the function's details.
}
\description{
This helper function is used within \code{mk_userdef_fn_dt} to process each unique
function definition in the input \code{data.table} and create a new row with the
parsed function information. It handles null functions by creating empty
entries and for non-null functions creates a character representation
and a callable function.
}
