% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_ep_for_targets.R
\name{group_ep_for_targets}
\alias{group_ep_for_targets}
\title{Group endpoints for targets pipeline}
\usage{
group_ep_for_targets(ep, n_per_group)
}
\arguments{
\item{ep}{A data.table containing the endpoint specification}

\item{n_per_group}{Number of endpoints per group.}
}
\value{
A data.table containing the endpoint specification with grouping
column.
}
\description{
Group endpoints for targets pipeline
}
