% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_mk_fn.R
\name{handle_mk_fn}
\alias{handle_mk_fn}
\title{Handle creation of endpoint def function}
\usage{
handle_mk_fn(
  fn,
  pipeline_id,
  r_functions_dir,
  type = c("mk_endpoint_def", "mk_criterion", "mk_adam"),
  env
)
}
\arguments{
\item{fn}{fn in list format}

\item{pipeline_id}{The pipeline ID}

\item{r_functions_dir}{The directory where the custom R scripts go}

\item{type}{Type of mk_* function: mk_endpoint_def (default), mk_criterion,
or mk_adam.}

\item{env}{Environment}
}
\description{
Handle creation of endpoint def function
}
\keyword{internal}
