% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_endpoints.R
\name{index_non_null_group_level}
\alias{index_non_null_group_level}
\title{Index Non-Null Group Levels}
\usage{
index_non_null_group_level(x)
}
\arguments{
\item{x}{A list representing group levels for endpoint grouping.}
}
\value{
A list containing only the non-null elements from the input list.
}
\description{
Takes a list representing group levels and returns a new list
with only the non-null elements. It is used to filter out null values from
a list of group levels, which can be necessary when defining the expanded
endpoint specifications.
}
