% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_filtered_endpoint_dt.R
\name{join_adam}
\alias{join_adam}
\title{Filter adam data}
\usage{
join_adam(
  ep,
  ep_fn_map,
  adam_db,
  filter_pop = TRUE,
  filter_period = TRUE,
  filter_trt = TRUE,
  filter_user_defined = TRUE
)
}
\arguments{
\item{adam_db}{data.table. A table containing the adam datasets associated
with each data_prepare}

\item{ep_unnest}{data.table. An unnested endpoint definition table}
}
\value{
a data.table with the filtered adam data.
}
\description{
Filter adam data
}
