% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_and_validate.R
\name{load_debug_session}
\alias{load_debug_session}
\title{Load Debugging Session}
\usage{
load_debug_session(debug_file)
}
\arguments{
\item{debug_file}{The path to the RDS file containing the debugging
environment.}
}
\value{
None; this function is used for its side effects.
}
\description{
Loads a previously staged debugging session from an RDS file
created by chef::with_stats_validation, setting up the environment to
debug the function call interactively.
}
