% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_function_inputs.R
\name{parse_function_input}
\alias{parse_function_input}
\title{Parse function inputs}
\usage{
parse_function_input(fn_input)
}
\arguments{
\item{fn_input}{A list representing the function input, which may contain a
function and its arguments.}
}
\value{
A function with arguments wrapped using \code{purrr::partial} if arguments
are provided; otherwise, returns the function unchanged.
}
\description{
This function processes a list representing a function input. If the function
is supplied with an argument, the argument should be specified in the form
\code{list(c(fun, arg=x))}. The function returns the input function unchanged if
no arguments are provided or wraps the function with its arguments using
\code{purrr::partial} if arguments are present.
}
