% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_stats.R
\name{prepare_for_stats}
\alias{prepare_for_stats}
\title{Prepare (and Expand) Endpoint for Statistical Analysis}
\usage{
prepare_for_stats(
  ep,
  analysis_data_container,
  fn_map,
  type = c("stat_by_strata_by_trt", "stat_by_strata_across_trt",
    "stat_across_strata_across_trt"),
  data_col = "dat",
  id_col = "strata_id"
)
}
\arguments{
\item{ep}{A \code{data.table} containing expanded endpoint definitions and
associated data, typically the output from \code{apply_criterion_by_strata}.}

\item{fn_map}{A \code{data.table} mapping endpoint definitions to statistical
functions.}

\item{type}{A character string specifying the type of statistics for which
the data is being prepared. Valid types are "stat_by_strata_by_trt",
"stat_by_strata_across_trt", and "stat_across_strata_across_trt".}

\item{data_col}{The name of the column in \code{ep} that contains the ADaM
dataset.}

\item{id_col}{The name of the column in \code{ep} that contains the unique
identifier for the strata.}
}
\value{
A \code{data.table} with expanded endpoints prepared for statistical
analysis.
}
\description{
Prepares the endpoint data for ingestion by the user-supplied
statistical function by expanding it based on the type of statistics to be
performed. It expands the data according to the specified statistical type,
which determines the grouping columns used for the expansion.
}
