% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_chef.R
\name{stage_pipeline}
\alias{stage_pipeline}
\title{Stage a targets pipeline so that you can work interactively with it}
\usage{
stage_pipeline(pipeline_id = NULL, pipeline_name = NULL)
}
\arguments{
\item{pipeline_id}{A character string. Must not contain non-alphanumeric
characters}

\item{pipeline_name}{Usually leave blank, only for special cases if you use a
custom naming convention for your pipelines (not recommended). If used,
leave \code{pipeline_id} blank and enter your custom pipeline name here.}
}
\value{
Nothing, run for side effects
}
\description{
To interact with a targets pipeline (e.g., run the pipeline,
load the completed targets from cache into memory), targets needs to know
which pipeline you want to work with. \verb{This function} is a thin wrapper
apound a Sys.setenv() call, and depends on the _targets.yaml file being set
up correctly (this happens automatically when you run use_pipeline() to set
up a new pipeline
}
