% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_and_validate.R
\name{try_and_validate}
\alias{try_and_validate}
\title{Try and Validate Wrapper for Statistical Functions}
\usage{
try_and_validate(
  expr_,
  expr_name = NA_character_,
  debug_dir = "debug",
  validator = function(expr_result) {
     NA_character_
 },
  stage_debugging = TRUE
)
}
\arguments{
\item{expr_}{The expression of type \code{call} to be evaluated, typically a call
to a statistical function.}

\item{expr_name}{The name of the expression, used for debugging purposes.}

\item{debug_dir}{The directory where debugging information will be stored.}

\item{validator}{A function used to validate the output of the expression.}

\item{stage_debugging}{A flag indicating whether to stage debugging
information in case of errors.}
}
\value{
The result of the evaluated expression if successful and valid.
}
\description{
Attempts to execute a statistical function and validates its
output. If the function fails or the output is invalid, it provides
meaningful error messages and sets up a debugging environment.
}
