% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_by_fns.R
\name{unnest_by_fns}
\alias{unnest_by_fns}
\title{Unnest data model by the supplied function columns}
\usage{
unnest_by_fns(dt, cols)
}
\arguments{
\item{dt}{A \code{data.table} object. The endpoint data model that contains nested function calls.}

\item{cols}{A character vector of quoted column names in \code{dt} that contain the nested function calls.}
}
\value{
A \code{data.table} in long format, where each row corresponds to a single function call
from the nested lists in the specified columns.
}
\description{
This function takes a \code{data.table} and a vector of column names. These columns are expected to
contain lists of function calls. The function 'unnests' these columns such that each function call
is represented on its own row in the resulting \code{data.table}. This is useful for processing or
analyzing the function calls separately.
}
