% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unnest_endpoint_functions.R
\name{unnest_endpoint_functions}
\alias{unnest_endpoint_functions}
\title{Unnest over user-supplied functions in endpoint definitions}
\usage{
unnest_endpoint_functions(
  endpoint_defs,
  fn_cols = c("data_prepare", "stat_by_strata_by_trt", "stat_by_strata_across_trt",
    "stat_across_strata_across_trt", "crit_endpoint", "crit_by_strata_by_trt",
    "crit_by_strata_across_trt"),
  env = parent.frame()
)
}
\arguments{
\item{endpoint_defs}{A \code{data.table} containing endpoint definitions with
nested function calls in the columns defined by the \code{fn_cols} argument.}

\item{fn_cols}{A character vector with default names of the columns that
contain the function calls. These columns are expected to be in the
\code{endpoint_defs} data.table.}

\item{env}{The environment from which to evaluate the functions. The default
is the parent frame from which the function is called.}
}
\value{
A \code{data.table} with each function call represented as a separate row,
including a unique hash for each function call.
}
\description{
This function processes a \code{data.table} containing endpoint definitions by
unnesting the function calls in specified columns. It is tailored for
endpoint definition specific data structure, where endpoint definitions
include user-supplied functions. After unnesting, it removes any rows where
the function is NULL (indicating an empty function) and computes a hash for
each function call for identification purposes.
}
