% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_and_validate.R
\name{validate_crit_output}
\alias{validate_crit_output}
\title{Validate Output from Criterion Functions}
\usage{
validate_crit_output(output)
}
\arguments{
\item{output}{The output from a criterion function.}
}
\value{
An error message if validation fails, otherwise NA.
}
\description{
Validates the output of criterion functions to ensure it is a
logical value. This is used to confirm that the criterion functions are
returning expected results.
}
