% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_endpoints_def.R
\name{validate_endpoints_def}
\alias{validate_endpoints_def}
\title{Validate user inputs to endpoint definition}
\usage{
validate_endpoints_def(endpoint_base)
}
\arguments{
\item{endpoint_base}{Endpoint definition data.table}
}
\value{
Run for side effects. If there is a validation error it stops the
program, otherwise the function returns a 'TRUE'
}
\description{
Validate user inputs to endpoint definition
}
