% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_usr_fn_args.R
\name{validate_usr_fn_args}
\alias{validate_usr_fn_args}
\title{Validate the user supplied functions arguments.}
\usage{
validate_usr_fn_args(
  fn,
  fn_type = c("data_prepare", "stat_by_strata_by_trt", "stat_by_strata_across_trt",
    "stat_across_strata_across_trt", "crit_endpoint", "crit_by_strata_by_trt",
    "crit_by_strata_across_trt"),
  fn_name = NA_character_
)
}
\arguments{
\item{fn}{(function) A non-primitive function}

\item{fn_type}{(character) giving the type of user def function.}

\item{fn_name}{(character) a custom name to use in error msg: will otherwise derive a name from function symbol.}
}
\value{
NA if function complies or throws an error.
}
\description{
Checks whether the expected function arguments are supplied.
Throws a meaningful error in case the expected and supplied arguments are mismatched.
}
