% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_event_index.R
\name{add_event_index}
\alias{add_event_index}
\title{Add Event Index to Expanded Endpoints}
\usage{
add_event_index(ep, analysis_data_container)
}
\arguments{
\item{ep}{A \code{data.table} that contains expanded endpoint definitions,
typically the output from \code{expand_over_endpoints}. It assumes the inclusion
of the columns \code{pop_var}, \code{pop_value}, \code{period_var}, \code{period_value},
\code{endpoint_filter}, \code{endpoint_group_filter}, and \code{custom_pop_filter}, which
are used to define the conditions for event indexing.}

\item{analysis_data_container}{A data.table containing the analysis data.}
}
\value{
A \code{data.table} similar to the input but with an additional
\code{event_index} column, which contains the indices of events as determined by
the specified conditions for each endpoint. The indices refer to the
\code{INDEX_} column in the clinical data. This \code{INDEX_} column is created by
chef when a user supplies a clinical dataset.
}
\description{
This function adds an event index to each row of the expanded
endpoints \code{data.table}. The event index is created based on variable-value
pairs and singleton conditions that define specific events of interest
within the dataset. This index can be used to identify events in the
user-supplied criteria and/or statistical functions
}
