% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_id.R
\name{add_id}
\alias{add_id}
\title{Add ID to user-defined endpoint groups}
\usage{
add_id(ep)
}
\arguments{
\item{ep}{A \code{data.table} containing endpoint definitions.}
}
\value{
data.table
}
\description{
Add ID to user-defined endpoint groups
}
