% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_criterion.R
\name{apply_criterion_endpoint}
\alias{apply_criterion_endpoint}
\title{Apply Endpoint Criterion}
\usage{
apply_criterion_endpoint(ep, analysis_data_container, fn_map)
}
\arguments{
\item{ep}{A \code{data.table} containing expanded endpoint definitions and
associated data, typically the output from \code{add_event_index}.}

\item{analysis_data_container}{data.table containing the analysis data.}

\item{fn_map}{A \code{data.table} mapping endpoint definitions to criterion
functions.}
}
\value{
A \code{data.table} with an additional logical column \code{crit_accept_endpoint}
indicating whether each endpoint meets the defined criteria.
}
\description{
Applies the user-supplied criteria functions to each endpoint to
determine eligibility based on the criteria. The result is a logical column
added to the data indicating whether each endpoint meets the criteria.
}
