% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_duplicate_functions.R
\name{check_duplicate_functions}
\alias{check_duplicate_functions}
\title{Check for duplicate function definitions}
\usage{
check_duplicate_functions(dir)
}
\arguments{
\item{dir}{The directory where the custom functions are defined}
}
\value{
run for side-effects, if multiple functions are encountered it throws
an error
}
\description{
Check for duplicate function definitions
}
