% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_endpoints.R
\name{define_expanded_ep}
\alias{define_expanded_ep}
\title{Define Expanded Endpoint Specifications}
\usage{
define_expanded_ep(x, group_by, col_prefix = "endpoint_group")
}
\arguments{
\item{x}{A \code{data.table} containing the data associated with the endpoints.}

\item{group_by}{A list specifying the grouping for endpoints, where
each element corresponds to a variable used for grouping endpoints and
contains the levels for that grouping variable.}

\item{col_prefix}{A prefix used to create the names of the metadata and
filter columns in the output \code{data.table}. Defaults to "endpoint_group".}
}
\value{
A \code{data.table} with additional columns for metadata and filter
conditions for each endpoint group. If the endpoint group is empty or
consists only of \code{NA} values, the function returns \code{NA}.
}
\description{
Takes a dataset and an endpoint group specification, and it
defines the expanded endpoint specifications. It creates metadata and
filter strings for each endpoint group, which are used to subset the data
accordingly. The function ensures that each endpoint group has the
necessary information for further analysis.
}
