% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_criteria.R
\name{evaluate_criteria}
\alias{evaluate_criteria}
\title{Evaluate criteria for inclusion of endpoints or endpoint strata}
\usage{
evaluate_criteria(
  endpoints,
  adam_set,
  criteria_type = c("endpoint", "subgroup_description", "subgroup_analysis")
)
}
\arguments{
\item{endpoints}{A data table with endpoint definitions.}

\item{adam_set}{A list of pre-processed ADaM tables.}

\item{criteria_type}{A string specifying the type of criteria. This can either
be "endpoint", "subgroup_description" or "subgroup_analysis", which apply
criteria on whether to include the endpoint, or by/across treatment and strata
in the final submission.}
}
\value{
A data table with endpoint definitions enriched with indications of
whether to keep the endpoint/strata or not.
}
\description{
Evaluate criteria for inclusion of endpoints or endpoint strata
}
