% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_endpoints.R
\name{expand_over_endpoints}
\alias{expand_over_endpoints}
\title{Expand endpoint definitions}
\usage{
expand_over_endpoints(ep, analysis_data_container)
}
\arguments{
\item{ep}{A \code{data.table} containing endpoint definitions, where each row
corresponds to a different endpoint and contains relevant attributes such
as the endpoint name, type, and criteria.}

\item{analysis_data_container}{data.table containing the analysis data.}
}
\value{
A \code{data.table} where each row corresponds to an expanded endpoint
definition
}
\description{
This function takes endpoint definitions and expands them based
on endpoint groups. Each row in the output corresponds to a single table or
figure list (TFL) definition, ensuring that each TFL has a unique
definition. The expansion is done by merging the endpoint definitions with
a mapping table that links these definitions to user-defined functions,
which are then applied to the pre-processed ADaM datasets to create the
expanded endpoints.
}
