% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{llist}
\alias{llist}
\title{Make list of lists}
\usage{
llist(...)
}
\arguments{
\item{...}{Elements to be included in the nested list}
}
\value{
A list object containing a list where each element is defined by
\code{...}
}
\description{
Make list of lists
}
