% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mk_userdef_fn_dt.R
\name{mk_userdef_fn_dt}
\alias{mk_userdef_fn_dt}
\title{Parse functions supplied by user}
\usage{
mk_userdef_fn_dt(x, env = parent.frame())
}
\arguments{
\item{x}{A \code{data.table} object representing the endpoint definition data
model, which should contain the following columns:
\itemize{
\item fn_type: Character vector specifying the type of function.
\item fn: List of callables (length 1) representing the function to be called.
\item fn_name: Character vector with the name of the function.
\item fn_hash: Character vector with a unique hash for each function.
}}

\item{env}{The environment from which to evaluate the functions. The default
is the parent frame from which the function is called.}
}
\value{
A \code{data.table} with each row containing a parsed function definition. The columns include:
\itemize{
\item fn_type: Character vector with the type of function.
\item fn_hash: Character vector with the unique hash of the function.
\item fn_name: Character vector with the name of the function.
\item fn_call_char: Character vector with the function call as a string.
\item fn_callable: List of callables representing the parsed function.
}
}
\description{
This function takes a \code{data.table} containing user-supplied function
definitions and parses them into a structured format. The input \code{data.table}
is expected to have specific columns that include the function type, the
function itself, the name of the function, and a unique hash for the
function. The output is a list of \code{data.table} objects, where each
\code{data.table} corresponds to a unique function type and contains detailed
information about each function, including a character representation and the
callable function itself.
}
