% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_chef.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run a {targets} pipeline}
\usage{
run_pipeline(pipeline_id = NULL, pipeline_name = NULL)
}
\arguments{
\item{pipeline_id}{A character string that will function as the pipeline ID.
Must not contain non-alphanumeric characters}

\item{pipeline_name}{Usually leave blank, only for special cases if you use a
custom naming convention for your pipelines (not recommended). If used,
leave \code{pipeline_id} blank and enter your custom pipeline name here.}
}
\value{
Nothing, run for side effects
}
\description{
A wrapper for targets::tar_make() that ensures the correct
pipeline is run, and the correct cache location is used for that pipeline.
Relies on the _targets.yaml file existing in the home directory of the
project.
}
